+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 1.0.0
+# 2.0.0
++ Refactor repository
++ Upgrade dependencies
++ Add level to getBackgroundColor
++ Add level to getColor
++ Refactor get message
+
+## 1.0.0
 + Upgrade dependencies
 
 ## 0.1.0